Clarify display language
The CRCC recommended displaying the COMMITMENT file in the root of the
repo, but also allowed for "an alternative" To make the proposal
concrete enough to adopt I removed the "or some alternative name"
language so it's specific.

I also changed the tense of the recommendation to other Pulp
repositories to present test instead of future. That way the merged PUP
will be that recommendation instead of a followon action.